requirements.txt
fastapi
uvicorn
pydantic>=2
python-dotenv
google-adk
httpx
mcp[cli]
google-genai